Update to the latest call object bundle URL construction logic, which was undone during a rebase.
