+++ src/module.js
@@ -1,5 +1,4 @@
 import EventEmitter from 'events';
-import Bowser from 'bowser';
 import { deepEqual } from 'fast-equals';
 import {
   filter,
@@ -84,7 +83,7 @@ import {
   DAILY_UI_REQUEST_FULLSCREEN,
   DAILY_UI_EXIT_FULLSCREEN,
 } from './shared-with-pluot-core/CommonIncludes.js';
-import { isReactNative } from './shared-with-pluot-core/Environment.js';
+import { isReactNative, browserInfo } from './shared-with-pluot-core/Environment.js';
 import WebMessageChannel from './shared-with-pluot-core/script-message-channels/WebMessageChannel';
 import ReactNativeMessageChannel from './shared-with-pluot-core/script-message-channels/ReactNativeMessageChannel';
 import CallObjectLoaderWeb from './call-object-loaders/CallObjectLoaderWeb.js';
@@ -206,58 +205,7 @@ export default class DailyIframe extends EventEmitter {
 
   static supportedBrowser() {
     methodNotSupportedInReactNative();
-    function supportsUnifiedPlanSDP(browser) {
-      return browser.satisfies({
-        electron: ">=6",
-        chromium: ">=75",
-        chrome: ">=75",
-        firefox: ">=67",
-        opera: ">=61",  // Corresponds to Chrome 75
-        // Technically Safari 12.1 supports Unified Plan SDP, but for simplicity
-        // we're just checking for 13.0.1 and above to avoid a 13.0.0 bug. 12.1
-        // will fail the isDisplayMediaAccessible() check anyway.
-        safari: ">=13.0.1",
-        edge: ">=79",   // Corresponds to Edgium
-      });
-    }
-
-    function isDisplayMediaAccessible() {
-      return navigator && navigator.mediaDevices && navigator.mediaDevices.getDisplayMedia;
-    }
-
-    const browser = Bowser.getParser(window.navigator.userAgent),
-          basic = browser.getBrowser(),
-          parsed = Bowser.parse(window.navigator.userAgent),
-          isValidBrowser = browser.satisfies({
-            electron: ">=6",
-            chromium: ">=61",
-            chrome: ">=61",
-            firefox: ">=63",
-            opera: ">=61",
-            safari: ">=12",
-            edge: ">=18",
-            iOS: {
-              chromium: "<0",
-              chrome: "<0",
-              firefox: "<0",
-              opera: "<0",
-              safari: ">=12",
-              edge: "<0",
-            }
-          }),
-        // See PluotUtil.isScreenSharingSupported() for a thorough explanation of this check
-        supportsScreenShare = !!(isValidBrowser && isDisplayMediaAccessible() && supportsUnifiedPlanSDP(browser)),
-        supportsSfu = !!(isValidBrowser && !browser.satisfies({edge: '<=18'}));
-
-    return {
-      supported: isValidBrowser,
-      mobile: parsed.platform.type === 'mobile',
-      name: basic.name,
-      version: basic.version,
-      supportsScreenShare,
-      supportsSfu,
-      // basic, parsed
-    };
+    return browserInfo();
   }
 
   //
